Fix FormationAddUnitPacket to assign next portId

+++ NebulaModel/Packets/Combat/GroundEnemy/DFGFormationAddUnitPacket.cs
@@ -8,12 +8,12 @@ public class DFGFormationAddUnitPacket
     {
         PlanetId = planetId;
         BaseId = baseId;
-        FormId = formId;
-        PortId = portId;
+        FormId = (byte)formId;
+        PortId = (ushort)portId;
     }
 
     public int PlanetId { get; set; }
     public int BaseId { get; set; }
-    public int FormId { get; set; }
-    public int PortId { get; set; }
+    public byte FormId { get; set; }
+    public ushort PortId { get; set; }
 }
