+++ NebulaModel/Packets/Combat/SpaceEnemy/DFSFormationAddUnitPacket.cs
@@ -7,11 +7,11 @@ public class DFSFormationAddUnitPacket
     public DFSFormationAddUnitPacket(int hiveAstroId, int formId, int portId)
     {
         HiveAstroId = hiveAstroId;
-        FormId = formId;
-        PortId = portId;
+        FormId = (byte)formId;
+        PortId = (ushort)portId;
     }
 
     public int HiveAstroId { get; set; }
-    public int FormId { get; set; }
-    public int PortId { get; set; }
+    public byte FormId { get; set; }
+    public ushort PortId { get; set; }
 }
